owsm_ctc_v4/speech_recognition/pytorch/requirements.txt
espnet
espnet_model_zoo
numpy
